requirements.txt
Django
graphene-django
django-graphql-jwt
openai
python-dotenv
django-filter